Initial code to perform phasing
Added a script to annotate the input VCF with information on variants
present in reference files. This is necessary because VCF files produced
by Long Ranger contain a mixture of phased and unphased haploid and
diploid calls which appears to cause the GATK's VariantAnnotator to
fail.

Added the code to identify mosaic variants in `main.py` and
`phasing.py`. An overview of mosaic variant identification is described
in the README. Code is currently slow, buggy and not well tested.